+++ triadics.go
@@ -16,8 +16,8 @@ func (ctx *Context) amend3(x, y, f V) V {
 	}
 }
 
-func (ctx *Context) amend3arrayI(x array, y int, f V) V {
-	if y < 0 || int(y) >= x.Len() {
+func (ctx *Context) amend3arrayI(x array, y int64, f V) V {
+	if y < 0 || y >= int64(x.Len()) {
 		return errf("@[x;y;f] : x out of bounds (%d)", y)
 	}
 	xy := x.at(int(y))
@@ -79,8 +79,8 @@ func (ctx *Context) amend4(x, y, f, z V) V {
 	}
 }
 
-func (ctx *Context) amend4arrayI(x array, y int, f, z V) V {
-	if y < 0 || int(y) >= x.Len() {
+func (ctx *Context) amend4arrayI(x array, y int64, f, z V) V {
+	if y < 0 || y >= int64(x.Len()) {
 		return errf("@[x;y;f;z] : x out of bounds (%d)", y)
 	}
 	xy := x.at(int(y))
